Moved legacytray package into central location and use it on uninstall.
