+++ internal/legacytray/legacytray_darwin.go
