+++ internal/legacytray/legacytray_windows.go
